+++ frame/support/Cargo.toml
@@ -11,7 +11,6 @@ version     = "2.6.8"
 
 [dependencies]
 # crates.io
-array-bytes           = { version = "1.3.3" }
 codec                 = { package = "parity-scale-codec", version = "2.1.1", default-features = false, features = ["derive"] }
 ethereum              = { version = "0.7.1", default-features = false, features = ["with-codec"] }
 impl-trait-for-tuples = { version = "0.2.1" }
