+++ frame/support/src/tests.rs
@@ -19,12 +19,7 @@
 // --- paritytech ---
 use sp_core::{H160, U256};
 // --- darwinia-network ---
-use crate::{
-	evm::IntoH160,
-	s2s::{RING_NAME, RING_SYMBOL},
-	*,
-};
-use array_bytes::{hex2array, hex2bytes};
+use crate::{evm::IntoH160, *};
 use std::str::FromStr;
 
 #[test]
@@ -45,25 +40,4 @@ fn test_into_dvm_account() {
 		(&b"longbytes..longbytes..longbytes..longbytes"[..]).into_h160(),
 		(&b"longbytes..longbytes"[..]).into_h160()
 	);
-}
-
-#[test]
-fn test_ring_symbol_encode() {
-	// Get this info: https://etherscan.io/address/0x9469d013805bffb7d3debe5e7839237e535ec483#readContract
-	let target_symbol = "0x52494e4700000000000000000000000000000000000000000000000000000000";
-	assert_eq!(to_bytes32(RING_SYMBOL), hex2array(target_symbol).unwrap());
-}
-
-#[test]
-fn test_ring_name_encode() {
-	// Get this info: https://etherscan.io/address/0x9469d013805bffb7d3debe5e7839237e535ec483#readContract
-	let target_name = "0x44617277696e6961204e6574776f726b204e617469766520546f6b656e000000";
-	assert_eq!(to_bytes32(RING_NAME), hex2array(target_name).unwrap());
-}
-
-#[test]
-fn test_ring_name_decode() {
-	let name = "0x44617277696e6961204e6574776f726b204e617469766520546f6b656e000000";
-	let raw = hex2bytes(name).unwrap();
-	assert_eq!(RING_NAME, &raw[..29]);
 }
