+++ primitives/s2s/Cargo.toml
@@ -10,6 +10,8 @@ repository  = "https://github.com/darwinia-network/darwinia-common/"
 version     = "2.6.8"
 
 [dependencies]
+# crates.io
+array-bytes = { version = "1.3.3" }
 # darwinia-network
 dp-asset    = { default-features = false, path = "../asset"}
 # paritytech
@@ -18,11 +20,15 @@ sp-core    = { default-features = false, git = "https://github.com/darwinia-netw
 sp-runtime = { default-features = false, git = "https://github.com/darwinia-network/substrate", tag = "darwinia-v0.11.5" }
 sp-std     = { default-features = false, git = "https://github.com/darwinia-network/substrate", tag = "darwinia-v0.11.5" }
 
+[dev-dependencies]
+dp-contract = { default-features = false, path = "../contract"}
+
 [features]
 default = ["std"]
 
 std = [
 	# darwinia-network
+	"dp-contract/std",
 	"dp-asset/std",
 	# paritytech
 	"codec/std",
