+++ primitives/s2s/src/lib.rs
@@ -18,42 +18,7 @@
 
 #![cfg_attr(not(feature = "std"), no_std)]
 
-mod test;
+pub mod call_params;
+pub mod token_info;
 
-// --- paritytech ---
-use sp_core::{H160, U256};
-use sp_std::{vec, vec::Vec};
-// --- darwinia-network ---
-use codec::{Decode, Encode};
-use dp_asset::token::TokenMetadata;
-use sp_runtime::AccountId32;
-
-/// The parameters box for the pallet runtime call.
-#[derive(Encode, Decode, Debug, PartialEq, Eq, Clone)]
-pub enum CallParams {
-	#[codec(index = 0)]
-	S2sIssuingPalletRegisterFromRemote(TokenMetadata),
-	#[codec(index = 1)]
-	S2sIssuingPalletIssueFromRemote(H160, U256, H160),
-	#[codec(index = 2)]
-	S2sBackingPalletUnlockFromRemote(H160, U256, AccountId32),
-}
-
-/// Creating a concrete message payload which would be relay to target chain.
-pub trait PayloadCreate<AccountId, MessagePayload>
-where
-	AccountId: Encode + Decode,
-{
-	fn encode_call(pallet_index: u8, call_params: CallParams) -> Result<Vec<u8>, &'static str> {
-		let mut encoded = vec![pallet_index];
-		encoded.append(&mut call_params.encode());
-		Ok(encoded)
-	}
-
-	fn payload(
-		submitter: AccountId,
-		spec_version: u32,
-		weight: u64,
-		call_params: CallParams,
-	) -> Result<MessagePayload, &'static str>;
-}
+pub use call_params::{CallParams, PayloadCreate};
